LayoutTests/svg/animations/animate-color-transparent-expected.txt

SVG 1.1 dynamic animation tests

Tests that 'transparent' is treated as a valid color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 127
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
